requirements.txt
# Vertex AI Pipelines
kfp==2.7.0
google-cloud-aiplatform==1.42.0

# BigQuery
google-cloud-bigquery==3.14.0

# Spanner
google-cloud-spanner==3.42.0

# その他
google-auth==2.27.0
